% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmca.R
\name{mcmca}
\alias{mcmca}
\title{Markov Chain Monte Carlo Objects for Arrays}
\usage{
mcmca(data = NA, start = 1, end = numeric(0), thin = 1)
}
\arguments{
\item{data}{an array of MCMC output}

\item{start}{the iteration number of the first observation}

\item{end}{the iteration number of the last observation}

\item{thin}{the thinning interval between consecutive observations}
}
\description{
Markov Chain Monte Carlo Objects for Arrays
}

