% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmca.R
\name{nvariables}
\alias{niterations}
\alias{nvariables}
\title{Dimensions of MCMC Array objects}
\usage{
nvariables(x)

niterations(x)
}
\arguments{
\item{x}{An mcmca object}
}
\description{
Dimensions of MCMC Array objects
}
\section{Functions}{
\itemize{
\item \code{niterations}: returns the number of iterations
}}

